map through habits db
